\name{configure_iaap}
\alias{configure_iaap}
\title{Procedure for setting path to IAAP-cli executable}
\usage{
configure_iaap()
}
\arguments{
}
\value{
returns path to IAAP-cli executable
}

\description{
 This procedure first checks whether the package home folder contains IAAP-cli executable. If not, then it guides the user to download the IAAP-cli archive and configure path to the archive. The procedure then decompresses the archive and sets up the path. 
}
\examples{
}