\name{create_from_frame}
\alias{create_from_frame}
\title{Internal function for converting between python structures}
\usage{
create_from_frame(df)
}
\arguments{
\item{df}{genotyping pandas dataframe from scbasic function}

}
\value{
Instance of class Data from Python SureTypeSC library with multiindex.

}
\description{
This functions is used internally in \code{scbasic} and converts python data frame in to SureTypeSC's data object.
}
\examples{
}
