\name{get_simpleind_df}
\alias{get_simpleind_df}
\title{Internal function that creates data frame with simple column index}
\usage{
get_simpleind_df(df)
}
\arguments{
\item{df}{data frame with column multiindex}
}
\value{
returns data frame with simple index
}

\description{
Internal function for index conversion. Function converts data frame's column multiindex into  a simple form. 
}
